Continue fiddling with type stuff
